+++ sdk/storage/azure-storage-file-share/src/main/java/com/azure/storage/file/share/ShareFileAsyncClient.java
@@ -53,7 +53,6 @@ import com.azure.storage.file.share.models.HandleItem;
 import com.azure.storage.file.share.models.NtfsFileAttributes;
 import com.azure.storage.file.share.models.PermissionCopyModeType;
 import com.azure.storage.file.share.models.Range;
-import com.azure.storage.file.share.models.ShareErrorCode;
 import com.azure.storage.file.share.models.ShareFileCopyInfo;
 import com.azure.storage.file.share.models.ShareFileDownloadAsyncResponse;
 import com.azure.storage.file.share.models.ShareFileDownloadHeaders;
@@ -258,32 +257,13 @@ public class ShareFileAsyncClient {
     Mono<Response<Boolean>> existsWithResponse(Context context) {
         return this.getPropertiesWithResponse(null, context)
             .map(cp -> (Response<Boolean>) new SimpleResponse<>(cp, true))
-            .onErrorResume(this::checkDoesNotExistStatusCode, t -> {
+            .onErrorResume(ModelHelper::checkDoesNotExistStatusCode, t -> {
                 HttpResponse response = t instanceof ShareStorageException
                     ? ((ShareStorageException) t).getResponse()
                     : ((HttpResponseException) t).getResponse();
                 return Mono.just(new SimpleResponse<>(response.getRequest(), response.getStatusCode(),
                     response.getHeaders(), false));
             });
-    }
-
-    private boolean checkDoesNotExistStatusCode(Throwable t) {
-        // ShareStorageException
-        return (t instanceof ShareStorageException
-            && ((ShareStorageException) t).getStatusCode() == 404
-            && (((ShareStorageException) t).getErrorCode() == ShareErrorCode.RESOURCE_NOT_FOUND
-                || ((ShareStorageException) t).getErrorCode() == ShareErrorCode.SHARE_NOT_FOUND))
-
-            /* HttpResponseException - file get properties is a head request so a body is not returned. Error
-             conversion logic does not properly handle errors that don't return XML. */
-            || (t instanceof HttpResponseException
-                && ((HttpResponseException) t).getResponse().getStatusCode() == 404
-                && (((HttpResponseException) t).getResponse()
-                    .getHeaderValue("x-ms-error-code")
-                    .equals(ShareErrorCode.RESOURCE_NOT_FOUND.toString())
-                    || (((HttpResponseException) t).getResponse()
-                        .getHeaderValue("x-ms-error-code")
-                        .equals(ShareErrorCode.SHARE_NOT_FOUND.toString()))));
     }
 
     /**