+++ sdk/storage/azure-storage-blob-cryptography/CHANGELOG.md
@@ -10,6 +10,19 @@
 
 ### Other Changes
 
+## 12.30.0 (2025-07-16)
+
+### Features Added
+- Added support for service version 2025-07-05.
+
+### Other Changes
+
+#### Dependency Updates
+- Upgraded `azure-storage-blob` from `12.30.0` to version `12.31.0`.
+- Upgraded `azure-json` from `1.4.0` to version `1.5.0`.
+- Upgraded `azure-core` from `1.55.2` to version `1.55.3`.
+- Upgraded `azure-core-http-netty` from `1.15.10` to version `1.15.11`.
+
 ## 12.31.0-beta.1 (2025-07-09)
 
 ### Features Added