+++ sdk/storage/azure-storage-internal-avro/CHANGELOG.md
@@ -10,6 +10,17 @@
 
 ### Other Changes
 
+## 12.16.0 (2025-07-16)
+
+### Features Added
+- Added support for service version 2025-07-05.
+
+### Other Changes
+
+#### Dependency Updates
+- Upgraded `azure-storage-common` from `12.29.0` to version `12.30.0`.
+- Upgraded `azure-core` from `1.55.2` to version `1.55.3`.
+
 ## 12.17.0-beta.1 (2025-07-09)
 
 ### Features Added